Added mode to i3wm, modified vimrc, and other stuff
@i3wm mode
Added a mode where I can move a workspace to another output. This is
very interesting if you hook a external display into you notebook and
want to move an entire workspace to this screen. Happens a lot to me.
This is very nice now.

@vimrc
Added pythonmode. This plugin comes VERY HANDY for python development.

@kitty conf
Moved it into its proper folder.

@i3blocks
I don't know what I have changed and I am actually to lazy to check it,
but I am not lazy to write this here.